Add basket model

+++ src/POS.Domain/ProductAggregate/Discount.cs
@@ -6,9 +6,9 @@ namespace POS.Domain.ProductAggregate
     public class Discount : BaseEntity<Guid>
     {
         public int Quantity { get; private set; }
-        public double DiscountPrice { get; private set; }
+        public decimal DiscountPrice { get; private set; }
 
-        public Discount(int quantity, double discountPrice)
+        public Discount(int quantity, decimal discountPrice)
         {
             this.Quantity = quantity;
             this.DiscountPrice = discountPrice;
@@ -19,7 +19,7 @@ namespace POS.Domain.ProductAggregate
             this.Quantity = quantity;
         }
 
-        public void SetDiscountPrice(double discountPrice)
+        public void SetDiscountPrice(decimal discountPrice)
         {
             this.DiscountPrice = discountPrice;
         }
